Fix small bugs and add readme

+++ plugin-core/src/main/java/cc/dreamcode/missions/mission/progress/MissionProgress.java
@@ -16,7 +16,7 @@ public class MissionProgress extends Document {
     private int missionId;
 
     @CustomKey("current-amount")
-    private int currentAmount;
+    private double currentAmount;
 
     @CustomKey("finished")
     private boolean finished;
